Eduardo

+++ src/main/java/com/example/Login/repository/LocalidadRepository.java
@@ -17,11 +17,35 @@ Integer>{
 	  
 	 Optional<Localidad> findByNombre(String nombre);
 	 
+//	 @Query(value = """
+//			    SELECT 
+//			       (v.localidad_destino_id) AS id,
+//			        l.nombre AS nombre,
+//			        COUNT(*)::BIGINT AS cantidad
+//			    FROM viaje v
+//			    JOIN compras c ON c.viaje_id = v.id
+//			    JOIN localidades l ON v.localidad_destino_id = l.id
+//			    WHERE c.estado_compra = 0
+//			    GROUP BY v.localidad_destino_id, l.nombre
+//			    ORDER BY cantidad DESC
+//			    LIMIT 10
+//			""", nativeQuery = true)
+	 
+	 
+//	 @Query(value = "SELECT v.localidad_destino_id AS id, l.nombre AS nombre, COUNT(*) AS cantidad " +
+//             "FROM viaje v " +
+//             "JOIN compras c ON c.viaje_id = v.id " +
+//             "JOIN localidades l ON v.localidad_destino_id = l.id " +
+//             "WHERE c.estado_compra = 0 " +
+//             "GROUP BY v.localidad_destino_id, l.nombre " +
+//             "ORDER BY cantidad DESC " +
+//             "LIMIT 10", nativeQuery = true)
+	 
 	 @Query(value = """
 			    SELECT 
-			       (v.localidad_destino_id) AS id,
-			        l.nombre AS nombre,
-			        COUNT(*)::BIGINT AS cantidad
+			       v.localidad_destino_id AS id,
+			       l.nombre AS nombre,
+			       COUNT(*) AS cantidad
 			    FROM viaje v
 			    JOIN compras c ON c.viaje_id = v.id
 			    JOIN localidades l ON v.localidad_destino_id = l.id
